requirements.txt
discord.py==2.5.2
httpx==0.28.1
openai==1.71.0
pydantic==2.11.3
pydantic_settings==2.8.1
python-dotenv==1.1.0